Added Disconnect Message
server.js: Added 'disconnect' message handling. Moved 'disconnect' and 'move' message handling back inside of socket.on('newplayer') because this made the compiler happy. Because we're dealing with asynchronous functions/callbacks, the compiler didn't know whether or not the socket.player object would be defined when we tried to access socket.player.id when the 'disconnect' message was outside of the 'newplayer' message hander.